Move util stuff to internal package
