+++ cmd/ensure.go
@@ -13,7 +13,7 @@ import (
 	"syscall"
 
 	"github.com/BurntSushi/toml"
-	"github.com/GetStream/vg/workspace"
+	"github.com/GetStream/vg/internal/workspace"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
