+++ cmd/initSettings.go
@@ -10,8 +10,8 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/GetStream/vg/utils"
-	"github.com/GetStream/vg/workspace"
+	"github.com/GetStream/vg/internal/utils"
+	"github.com/GetStream/vg/internal/workspace"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
