+++ cmd/localInstall.go
@@ -7,8 +7,8 @@ import (
 	"fmt"
 	"path/filepath"
 
-	"github.com/GetStream/vg/utils"
-	"github.com/GetStream/vg/workspace"
+	"github.com/GetStream/vg/internal/utils"
+	"github.com/GetStream/vg/internal/workspace"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
