+++ cmd/querySettings.go
@@ -6,7 +6,7 @@ package cmd
 import (
 	"fmt"
 
-	"github.com/GetStream/vg/workspace"
+	"github.com/GetStream/vg/internal/workspace"
 	"github.com/spf13/cobra"
 )
