+++ cmd/uninstall.go
@@ -8,8 +8,8 @@ import (
 	"os"
 	"path/filepath"
 
-	"github.com/GetStream/vg/utils"
-	"github.com/GetStream/vg/workspace"
+	"github.com/GetStream/vg/internal/utils"
+	"github.com/GetStream/vg/internal/workspace"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 )
